Move the AutoType tests to tests
