+++ tests/autotype/AutoTypeTest.cpp
